+++ src/internal.rs
@@ -3,3 +3,5 @@ pub trait Sealed {}
 impl<T, E> Sealed for Result<T, E> {}
 
 impl<T> Sealed for Option<T> {}
+
+impl Sealed for bool {}